restrict placement of progenitors, move coord functions to new file
